+++ tests/test.sh
@@ -63,5 +63,15 @@ test_get_python_version() {
     assertSame "$expected" "$actual"
 }
 
+test_python_interpreter_set_incorrectly() {
+    return_to="$(pwd)"
+    cd "$WORKON_HOME"
+    mkvirtualenv --no-site-packages no_wrappers
+    output=`VIRTUALENVWRAPPER_PYTHON=$VIRTUAL_ENV/bin/python $SHELL $return_to/virtualenvwrapper.sh 2>&1`
+    assertTrue "Unexpected message: $output" "echo \"$output\" | grep 'Could not find Python module virtualenvwrapper.hook_loader'"
+    assertFalse "Failed to detect invalid Python location" "VIRTUALENVWRAPPER_PYTHON=$VIRTUAL_ENV/bin/python $SHELL $return_to/virtualenvwrapper.sh >/dev/null 2>&1"
+    cd "$return_to"
+    deactivate
+}
 
 . "$test_dir/shunit2"