+++ docs/source/history.rst
@@ -7,6 +7,9 @@ Dev
   - Add support for ksh.  Thanks to Doug Latornell for doing the
     research on what needed to be changed.
   - Switch to ``tempfile`` command for creating temporary hook files.
+  - Test import of virtualenvwrapper.hook_loader on startup and report
+    the error in a way that should help the user figure out how to fix
+    it.
 
 2.0.2